Fixed bug 1641 - avoid allocating nsstring from char*
Vittorio Giovara 2012-11-12 05:52:47 PST

Changesets 4f272256d172 and 42214b6959c5 introduce two neat features for
logging and alertbox on ios and osx.
However the NSString allocated (and a few other objects) are not freed by the
autorelease pool when created by +alloc and -initWithStuff: and this will
create leaks. While negligible on osx, on mobile it's better not to have leaks.
Attached is a patch that should take care of the problems on both platforms.