wayland: Determine GL ES context version at runtime
Use SDL_GL_CONTEXT_MAJOR_VERSION when setting up EGL and
the rendering context. This allows the library user to select
whether OpenGL ES 1.1 or OpenGL ES 2.0 is used for rendering
when they only use SDL for bringing up an OpenGL (ES) window.

By default, SDL_GL_CONTEXT_MAJOR_VERSION is set to the
right version (in src/video/SDL_video.c) depending on which
renderer was selected, so the default behavior will be the
same when the user doesn't set a custom context version.